+++ docs/CHANGELOG.md
@@ -1,3 +1,10 @@
+## [v0.5.0](https://github.com/jai-python3/validation-component-bootstrap-utils/tree/v0.5.0) - 2024-03-27
+
+ADDS:
+- support for custom validation functions via configuration file
+- support for tab-delimited files with .txt filename extension
+
+
 ## [v0.3.0](https://github.com/jai-python3/validation-component-bootstrap-utils/tree/v0.3.0) - 2024-02-05
 
 ADDS: